Remove trademark violation
Chia Network owns the registered US Trademark 6,428,600 and the registered EU Trademark 017930981 on the word "chia" as applied to cryptocurrencies as well as many other jurisdictions around the world. This developer (@snight) is free to name it something like greenrose or coinrose, but the current use is a violation of our Marks.

We have already taken over the chiarose.com domain name in ICANN arbitration.